Integrating MQTT. Addeda class holding core smart sensor and all controllers/observers - simplifying main.

+++ include/Sensors/CO2Observer.h
@@ -8,5 +8,5 @@ public:
 
     /// @brief Inform of a change in CO2 level (in total parts-per-million).
     /// @param present Returns CO2 in ppm. 
-    virtual void onCO2(uint16_t co2) = 0;
+    virtual void onCO2(float co2) = 0;
 };
