+++ include/Sensors/TVOCObserver.h
@@ -8,5 +8,5 @@ public:
 
     /// @brief Inform of a change in TVOC level (in parts per billion).
     /// @param present Returns total TVOC in ppb,
-    virtual void onTVOC(uint16_t tvoc) = 0;
+    virtual void onTVOC(float tvoc) = 0;
 };
